feat: extract updatecli code to share
